+++ lib/any_cache/adapters/active_support_mem_cache_store.rb
@@ -2,48 +2,54 @@
 
 module AnyCache::Adapters
   # @api private
-  # @since 0.1.0
-  class ActiveSupportRedisCacheStore < Basic
-    # TODO: think about locks (Concurrent::ReentrantReadWriteLock)
+  # @since 0.2.0
+  class ActiveSupportMemCacheStore < Basic
+    # NOTE: think about locks (Concurrent::ReentrantReadWriteLock)
 
     class << self
       # @param driver [Object]
-      # @return [Boolean]
+      # @retunr [Boolean]
       #
       # @api private
-      # @since 0.1.0
+      # @since 0.2.0
       def supported_driver?(driver)
-        AnyCache::Drivers::ActiveSupportRedisCacheStore.supported_source?(driver)
+        AnyCache::Drivers::ActiveSupportMemCacheStore.supported_source?(driver)
       end
     end
-
-    # @since 0.1.0
-    def_delegators :driver, :delete, :clear
-
-    # @return [NilClass]
-    #
-    # @api private
-    # @since 0.1.0
-    NO_EXPIRATION_TTL = nil
 
     # @return [Integer]
     #
     # @api private
-    # @since 0.1.0
+    # @since 0.2.0
+    NO_EXPIRATION_TTL = nil
+
+    # @return [NilClass]
+    #
+    # @api private
+    # @since 0.2.0
     DEAD_TTL = 0
 
     # @return [Integer]
     #
     # @api private
-    # @since 0.1.0
+    # @since 0.2.0
     DEFAULT_INCR_DECR_AMOUNT = 1
+
+    # @return [Integer]
+    #
+    # @api private
+    # @since 0.2.0
+    INITIAL_DECREMNETED_VALUE = 0
+
+    # @since 0.2.0
+    def_delegators :driver, :delete, :clear
 
     # @param key
     # @param options [Hash]
     # @return [Object]
     #
     # @api private
-    # @since 0.1.0
+    # @since 0.2.0
     def read(key, **options)
       driver.read(key, raw: true)
     end
@@ -54,7 +60,7 @@ module AnyCache::Adapters
     # @return [void]
     #
     # @api private
-    # @since 0.1.0
+    # @sicne 0.2.0
     def write(key, value, **options)
       expires_in = options.fetch(:expires_in, NO_EXPIRATION_TTL)
 
@@ -62,12 +68,12 @@ module AnyCache::Adapters
     end
 
     # @param key [String]
-    # @param amount [Integer, Float]
-    # @option expires_in [Integer]
+    # @param amount [Integer]
+    # @options expires_in [Integer]
     # @return [Integer, Float]
     #
     # @api private
-    # @since 0.1.0
+    # @since 0.2.0
     def increment(key, amount = DEFAULT_INCR_DECR_AMOUNT, **options)
       expires_in = options.fetch(:expires_in, NO_EXPIRATION_TTL)
       is_initial = !read(key) # TODO: rewrite with #exist?(key)
@@ -82,18 +88,22 @@ module AnyCache::Adapters
     end
 
     # @param key [String]
-    # @param amount [Integer, Float]
+    # @param amount [Integer]
     # @option expires_in [Integer]
     # @return [Integer, Float]
     #
     # @api private
-    # @since 0.1.0
+    # @since 0.2.0
     def decrement(key, amount = DEFAULT_INCR_DECR_AMOUNT, **options)
       expires_in = options.fetch(:expires_in, NO_EXPIRATION_TTL)
       is_initial = !read(key) # TODO: rewrite with #exist?(key)
 
       if is_initial
-        write(key, -amount, expires_in: expires_in) && -amount
+        # NOTE: Dalli::Client can't decrement:
+        #   - non-raw values;
+        #   - values lower than zero;
+        #   - empty entries;
+        write(key, INITIAL_DECREMNETED_VALUE, expires_in: expires_in) && INITIAL_DECREMNETED_VALUE
       else
         driver.decrement(key, amount).tap do
           expire(key, expires_in: expires_in) if expires_in
@@ -106,7 +116,7 @@ module AnyCache::Adapters
     # @return [void]
     #
     # @api private
-    # @since 0.1.0
+    # @since 0.2.0
     def expire(key, expires_in: DEAD_TTL)
       read(key).tap do |value|
         is_alive = expires_in ? expires_in.positive? : false
@@ -119,7 +129,7 @@ module AnyCache::Adapters
     # @return [void]
     #
     # @api private
-    # @since 0.1.0
+    # @sicne 0.2.0
     def persist(key, **options)
       read(key).tap { |value| write(key, value) }
     end