+++ lib/any_cache/drivers/dalli.rb
@@ -14,12 +14,12 @@ module AnyCache::Drivers::Dalli
     end
 
     # @param settings [Qonfig::Settings]
-    # @return [::Redis]
+    # @return [::Dalli::Client]
     #
     # @api private
     # @sicne 0.2.0
     def build(settings)
-      ::Dalli::Client.new(settings.servers, settings.options)
+      ::Dalli::Client.new(Array(settings.servers), settings.options)
     end
   end
 end