+++ spec/features/increment_spec.rb
@@ -122,6 +122,7 @@ describe 'Operation: #increment' do
       it 'creates new permanent entry with corresponding initial value' do
         ini_value = rand(1..100)
         new_amount = cache_store.increment(entry[:key], ini_value)
+
         expect(new_amount).to eq(ini_value)
 
         sleep(expiration_time + 1) # NOTE: try to expire current entry